Tests and parser error recovery fully working

+++ test/rs/ac/bg/etf/pp1/MJParserTest.java
@@ -35,29 +35,34 @@ public class MJParserTest {
 	        Symbol s = p.parse();  //pocetak parsiranja
 	        
 	        Program prog = (Program)(s.value);
+			if(!p.errorDetected){
+				TabDerived.init();
+				// ispis sintaksnog stabla
+				log.info(prog.toString(""));
+				log.info("===================================");
 
-			TabDerived.init();
-			// ispis sintaksnog stabla
-			log.info(prog.toString(""));
-			log.info("===================================");
+				// ispis prepoznatih programskih konstrukcija
+				SemanticPass v = new SemanticPass();
+				prog.traverseBottomUp(v);
 
-			// ispis prepoznatih programskih konstrukcija
-			SemanticPass v = new SemanticPass();
-			prog.traverseBottomUp(v);
+				log.info("===================================");
+				TabDerived.dump();
+				if(!p.errorDetected && v.passed()){
+					File objFile = new File("test/program.obj");
+					if(objFile.exists()) objFile.delete();
 
-			log.info("===================================");
-			TabDerived.dump();
-			if(!p.errorDetected && v.passed()){
-				File objFile = new File("test/program.obj");
-				if(objFile.exists()) objFile.delete();
-
-				CodeGenerator codeGenerator = new CodeGenerator();
-				prog.traverseBottomUp(codeGenerator);
-				Code.dataSize = v.nVars;
-				Code.mainPc = codeGenerator.getMainPc();
-				Code.write(new FileOutputStream(objFile));
-				log.info("Parsiranje uspesno zavrseno!");
-			}else{
+					CodeGenerator codeGenerator = new CodeGenerator();
+					prog.traverseBottomUp(codeGenerator);
+					Code.dataSize = v.nVars;
+					Code.mainPc = codeGenerator.getMainPc();
+					Code.write(new FileOutputStream(objFile));
+					log.info("Parsiranje uspesno zavrseno!");
+				}
+				else{
+					log.error("Parsiranje nije uspesno zavrseno!");
+				}
+			}
+			else{
 				log.error("Parsiranje nije uspesno zavrseno!");
 			}
